gemini-tts-mcp/tts_api/requirements.txt
fastapi
uvicorn
google-cloud-texttospeech>=2.29.0
